Make subindex max size and entries configable
git-svn-id: file:///home/toad/git-migration/temprepository/trunk/plugins/XMLSpider@26085 67a373e5-eb02-0410-a15c-ee090a768436

+++ src/plugins/XMLSpider/db/Config.java
@@ -14,7 +14,7 @@ public class Config extends Persistent implements Cloneable {
 	 */
 	private String indexDir;
 	private int indexMaxEntries;
-	private long indexSubindexMaxSize;
+	private int indexSubindexMaxSize;
 
 	private String indexTitle;
 	private String indexOwner;
@@ -87,12 +87,12 @@ public class Config extends Persistent implements Cloneable {
 		return indexMaxEntries;
 	}
 
-	public synchronized void setIndexSubindexMaxSize(long indexSubindexMaxSize) {
+	public synchronized void setIndexSubindexMaxSize(int indexSubindexMaxSize) {
 		assert !isPersistent();
 		this.indexSubindexMaxSize = indexSubindexMaxSize;
 	}
 
-	public synchronized long getIndexSubindexMaxSize() {
+	public synchronized int getIndexSubindexMaxSize() {
 		return indexSubindexMaxSize;
 	}
